Move some testutils functions to utils_linux.go
